+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Changed proposal snapshot type from string to integer
+
 ## [0.4.1] - 2024-03-19
 
 ### Added